requirements.txt
together==1.3.3
python-dotenv==1.0.1
langchain-together==0.2.0
langchain==0.3.7
langchain-community==0.3.5
faiss-cpu==1.9.0
chromadb==0.5.18
colorlog==6.9.0
langchain-chroma==0.1.4